Project reordering

+++ include/sys_utils.h
@@ -115,8 +115,8 @@ unsigned int strlen(const char * s);
 int     sprintf(char * buf, const char *fmt, ...);
 int     sys_sscanf(const char * buf, const char * fmt, ...);
 unsigned long simple_strtoul(const char *cp,char **endp,unsigned int base);
-int     sys_checksum_ok(uchar * ptr, int len);
-uint    sys_checksum(uchar * ptr, int len);
+int     sys_checksum_ok(ushort * ptr, int len);
+uint    sys_checksum(ushort * ptr, int len);
